Adjusted to only dispaly instructions when viewing on mobile display
